Extract components & hocs
